optimize_dependencies by default
This will improve build parallelism significantly, but at the cost that cmake must be aware of generated srcs.  This is automatic for e.g. configure file, protobuf, etc.  
However if you use e.g. exec_program(generate_sources.py), then the targets which depend on the target with these sources must be marked as dependent.